+ Added slug for better url-friendly + Fixed checkout did not calculate price correctly + Fixed checkout did not decrease quantity in stock + Renamed some stuff ~ ~
+ Improved UI

+++ back-end/src/main/java/com/ring/bookstore/model/OrderDetail.java
@@ -33,7 +33,7 @@ public class OrderDetail {
     private Long id;
 
     @Column
-    private Integer amount;
+    private Short amount;
 
     @Column
     private Double price;
